디테일페이지 사진보이기, 사진문자열제거

+++ src/main/java/net/skhu/domain/repository/CommunityCommentRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import net.skhu.domain.entity.CommunityComment;
 
-public interface CommunityCommentRepository extends JpaRepository<CommunityComment, Integer> {
+public interface CommunityCommentRepository extends JpaRepository<CommunityComment, Long> {
 }
